pyproject.toml
[tool.poetry]
name = "ssl-models"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = "^2.2.2"
pytorch-lightning = "^2.2.1"
vit-pytorch = "^1.6.5"
einops = "^0.7.0"
jsonargparse = {extras = ["signatures"], version = "^4.27.7"}
geopandas = "^0.14.3"
rasterio = "^1.3.9"
ipython = "^8.23.0"
requests = "^2.31.0"
lightning = "^2.2.2"
tensorboard = "^2.16.2"
segmentation-models-pytorch = "^0.3.3"
omegaconf = "^2.3.0"
clearml = "^1.15.1"
python-dotenv = "^1.0.1"
hydra-zen = "^0.13.0"
h5py = "^3.12.1"
transformers = "^4.44.2"
torchgeo = "^0.6.0"
seaborn = "0.13.2"
opencv-python = "^4.11.0.86"
albumentations = "1.4.0"
pyogrio = "0.10.0"
pyarrow = "18.1.0"
matplotlib = "^3.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
